requirements.txt
Django>=3.0.4,<3.1.0
djangorestframework>=3.11.0,<3.12.0
django-cors-headers>=3.2.1,<3.3.0
psycopg2>=2.8.4,<2.9.0
flake8>=3.7.9,<3.8.0
dj-database-url>=0.3.0,<0.4.0
python-decouple>=3.4.0,<4.0.0
django-rest-knox>=4.1.0,<4.2.0
PyYAML>=5.4.1,<5.5.0
uritemplate>=3.0.1.<3.1.0
coreapi>=2.3.3,<2.4.0
schedule>=1.1.0,<1.2.0
pytest>=6.2.5,<6.3.0
pytest-django>=4.4.0,<4.5.0
mixer>=7.2.0,<7.3.0
pytest-cov>=3.0.0,<3.1.0
pytest-xdist>=2.4.0,<2.5.0
black==21.9b0